+++ src/com/nixinova/main/Mineo.java
@@ -12,7 +12,7 @@ import com.nixinova.world.World;
 
 public class Mineo {
 
-	public static final String VERSION = "0.0.10_7";
+	public static final String VERSION = "0.0.10_8";
 	public static final String TITLE = "Mineo " + VERSION;
 
 	public static JFrame frame;